Refactor to make the registry of columns more efficient, do not re-use the ember-tabular-columns, create a separate component to handle this
Reference issue #88
