+++ addon/components/ember-tabular-column.js
@@ -1,9 +1,8 @@
 import Ember from 'ember';
 import layout from 'ember-tabular/templates/components/ember-tabular-column';
-import EmberTabularHelpers from 'ember-tabular/mixins/components/ember-tabular-helpers';
 
 /**
-* Generates a column object used within ember tabular table headings
+* Generates a column object.
 * Sets sensible defaults which can be overwritten when setting up the template
 * for ember tabular
 *
@@ -28,7 +27,7 @@ import EmberTabularHelpers from 'ember-tabular/mixins/components/ember-tabular-h
 *
 * @class EmberTabularColumn
 */
-export default Ember.Component.extend(EmberTabularHelpers, {
+export default Ember.Component.extend({
   layout,
   /**
   * @property tagName
@@ -37,7 +36,6 @@ export default Ember.Component.extend(EmberTabularHelpers, {
   */
   tagName: '',
   action: null,
-  registry: null,
   hasProperty: Ember.computed('property', 'column', function () {
     const property = this.get('property');
     const column = this.get('column');
@@ -46,40 +44,4 @@ export default Ember.Component.extend(EmberTabularHelpers, {
     }
     return false;
   }),
-  init() {
-    this._super(...arguments);
-    const registry = this.get('registry');
-    const property = this.get('property');
-    const selectable = this.get('selectable');
-    const label = this.get('label') || this._formatColumnLabel(property);
-    const isActive = this._checkIfUndefined(this.get('isActive'), true);
-    const isCustom = this.get('isCustom') || false;
-    const filter = this._checkIfUndefined(this.get('filter'), true);
-    const list = this.get('list') || null;
-    const sort = this._checkIfUndefined(this.get('sort'), true);
-    const type = this.get('type') || 'text';
-    let column = {
-      property: property,
-      selectable: selectable,
-      label: label,
-      isActive: isActive,
-      isCustomTemplate: isCustom,
-      filter: filter,
-      list: list,
-      sort: sort,
-      type: type,
-    };
-    if (typeof this.get('isCustom') !== 'undefined' && registry) {
-      let item = registry.find((el) => {
-        return el.property === property;
-      });
-      if (!item) {
-        // pass action up to ember-tabular to add column to registry
-        this.get('addToRegistry')(column);
-      }
-    }
-  },
-  _checkIfUndefined(property, defaultValue) {
-    return typeof property !== 'undefined' ? property : defaultValue;
-  },
 });
