+++ app/components/ember-tabular-column-register.js
@@ -1,0 +1,3 @@
+import EmberTabularColumnRegister from 'ember-tabular/components/ember-tabular-column-register';
+
+export default EmberTabularColumnRegister;
